+++ config/initializers/categories.rb
@@ -1,4 +1,0 @@
-hashes = YAML.load_file(Rails.root.join('config/categories.yml'))
-STATIC_CATEGORIES = hashes.collect do |hash|
-  OpenStruct.new hash
-end