+++ config/initializers/static_data.rb
@@ -1,0 +1,12 @@
+raw_data = YAML.load_file(Rails.root.join('config/static_data.yml'))
+
+STATIC_DATA = {
+  categories: raw_data['categories'].collect { |hash| OpenStruct.new hash },
+  department_topics: raw_data['department_topics'].collect {|k, v|
+    [k, OpenStruct.new({
+      blurb: v['blurb'],
+      topics: v['topics'].collect {|raw_topic|
+        OpenStruct.new raw_topic
+      }})]
+  }.to_h
+}