+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This file is influenced by http://keepachangelog.com/.
 
 - Support setting user names in admin
 - Allow users to specify their name when they sign up
+- Static related topics for department homepages
 
 ### Changed